+++ cookbooks/python/attributes/default.rb
@@ -21,22 +21,19 @@
 default['python']['install_method'] = 'package'
 
 if python['install_method'] == 'package'
-  default['python']['prefix_dir']         = '/usr'
+  case platform
+  when "smartos"
+    default['python']['prefix_dir']         = '/opt/local'
+  else
+    default['python']['prefix_dir']         = '/usr'
+  end
 else
   default['python']['prefix_dir']         = '/usr/local'
 end
 
-case platform
-when "redhat","centos","scientific","fedora","suse","amazon"
-  set['python']['pip']['prefix_dir']  = '/usr'
-when "debian","ubuntu"
-  set['python']['pip']['prefix_dir']  = '/usr/local'
-else
-  set['python']['pip']['prefix_dir']  = '/usr/local'
-end
+default['python']['binary'] = "#{python['prefix_dir']}/bin/python"
 
 default['python']['url'] = 'http://www.python.org/ftp/python'
-default['python']['version'] = '2.7.1'
+default['python']['version'] = '3.3.1'
 default['python']['checksum'] = '80e387bcf57eae8ce26726753584fd63e060ec11682d1145af921e85fd612292'
 default['python']['configure_options'] = %W{--prefix=#{python['prefix_dir']}}
-