+++ cookbooks/python/metadata.rb
@@ -1,8 +1,9 @@
+name              "python"
 maintainer        "Opscode, Inc."
 maintainer_email  "cookbooks@opscode.com"
 license           "Apache 2.0"
 description       "Installs Python, pip and virtualenv. Includes LWRPs for managing Python packages with `pip` and `virtualenv` isolated Python environments."
-version           "1.0.8"
+version           "1.2.2"
 
 depends           "build-essential"
 depends           "yum"
@@ -13,6 +14,6 @@ recipe "python::source", "Installs python from source."
 recipe "python::pip", "Installs pip from source."
 recipe "python::virtualenv", "Installs virtualenv using the python_pip resource."
 
-%w{ debian ubuntu centos redhat fedora freebsd }.each do |os|
+%w{ debian ubuntu centos redhat fedora freebsd smartos }.each do |os|
   supports os
 end